Refactoring: Delete old 'render.js' and apply a new 'render.js' from framework
