+++ src/presenter/popup-presenter.js
@@ -1,5 +1,5 @@
 import { isEscape } from '../util.js';
-import {render} from '../render.js';
+import {render} from '../framework/render.js';
 import PopupView from '../view/popup-view.js';
 
 export default class PopupPresenter {
